Move modules around to place them in (seemingly) appropriate locations
