+++ src/mac/frame/security_control.rs
@@ -1,63 +1,8 @@
-//! All of the fields and their contents that are used in the
-//! Auxiliary Security header can be found in this module.
-//!
-//! All of the below data structures are defined in section 7.4 of the 802.15.4-2011 standard
+//! The security control header is described here
 
-use byte::{check_len, BytesExt, TryRead, LE};
+use byte::{BytesExt, TryRead, LE};
 
-use crate::mac::frame::DecodeError;
-
-/// A struct describing the Auxiliary Security Header
-/// See: section 7.4 of the 802.15.4-2011 standard
-#[derive(Clone, Copy, Debug, Eq, Hash, PartialEq)]
-pub struct AuxiliarySecurityHeader {
-    /// The control field in the Auxiliary Security Header
-    pub control: SecurityControl,
-    /// If the frame_counter_suppression field in [`AuxiliarySecurityHeader::control`] is not set, this field contains
-    /// the frame counter, otherwise it is None
-    pub frame_counter: u32,
-    /// If the key_identifier field in [`AuxiliarySecurityHeader::control`] is set, this field contains the key identifier
-    /// of this frame, otherwise it is None
-    pub key_identifier: Option<KeyIdentifier>,
-}
-
-impl TryRead<'_> for AuxiliarySecurityHeader {
-    fn try_read(bytes: &[u8], _ctx: ()) -> byte::Result<(Self, usize)> {
-        let offset = &mut 0;
-        check_len(bytes, 1)?;
-
-        let control: SecurityControl = bytes.read(offset)?;
-        let frame_counter = bytes.read_with(offset, LE)?;
-
-        let mut header = AuxiliarySecurityHeader {
-            control,
-            frame_counter,
-            key_identifier: None,
-        };
-
-        // Construct the KeyIdentifier in case the KeyIdentifierMode is not None
-        if control.key_id_mode != KeyIdentifierMode::None {
-            let mut key_identifier = KeyIdentifier {
-                key_source: None,
-                key_index: 0,
-            };
-            match control.key_id_mode {
-                KeyIdentifierMode::KeySource4 => {
-                    key_identifier.key_source = Some(KeySource::Short(bytes.read(offset)?));
-                }
-                KeyIdentifierMode::KeySource8 => {
-                    key_identifier.key_source = Some(KeySource::Long(bytes.read(offset)?));
-                }
-                _ => {}
-            }
-            let key_index = bytes.read(offset)?;
-            key_identifier.key_index = key_index;
-            header.key_identifier = Some(key_identifier);
-        }
-
-        Ok((header, *offset))
-    }
-}
+use super::DecodeError;
 
 /// The Security Control header
 ///
@@ -159,22 +104,6 @@ impl KeyIdentifierMode {
             _ => None,
         }
     }
-}
-
-/// A key identifier, comprised of a key index and possibly a key source
-#[derive(Clone, Copy, Debug, Eq, Hash, PartialEq)]
-pub struct KeyIdentifier {
-    key_source: Option<KeySource>,
-    key_index: u8,
-}
-
-/// A key source
-#[derive(Clone, Copy, Debug, Eq, Hash, PartialEq)]
-pub enum KeySource {
-    /// A key source that is 4 octets long
-    Short(u32),
-    /// A key source that is 8 octets long
-    Long(u64),
 }
 
 mod mask {
