--- src/mac/security/mod.rs
@@ -1,5 +1,0 @@
-//! Security of 802.15.4 frames
-//!
-//! See section 7 of the 802.15.4-2011 standard
-
-pub mod auxiliary_security_header;
